+++ testData/org/elixir_lang/model/psi/rename/matrix/function_defdelegate_head_after.ex
@@ -1,0 +1,7 @@
+defmodule DelegateTarget do
+  def renamee(x), do: x
+end
+
+defmodule Delegator do
+  defdelegate fresh(x), to: DelegateTarget, as: :renamee
+end